Testing

+++ webb_grp_imgs.py
@@ -196,12 +196,12 @@ def make_spline_img_3d(pos, Ndim, tree, ls, smooth, f, oversample,
         temp_img = np.zeros((Ndim, Ndim), dtype=np.float64)
 
         # Define x and y positions of pixels
-        X, Y, Z = np.meshgrid(np.arange(0, Ndim, 1, dtype=np.int16),
-                              np.arange(0, Ndim, 1, dtype=np.int16),
-                              np.arange(0, Ndim, 1, dtype=np.int16))
+        X, Y, Z = np.meshgrid(np.arange(0, Ndim, 1, dtype=np.int32),
+                              np.arange(0, Ndim, 1, dtype=np.int32),
+                              np.arange(0, Ndim, 1, dtype=np.int32))
 
         # Define pixel position array for the KDTree
-        pix_pos = np.zeros((X.size, 3), dtype=np.int16)
+        pix_pos = np.zeros((X.size, 3), dtype=np.int32)
         pix_pos[:, 0] = X.ravel()
         pix_pos[:, 1] = Y.ravel()
         pix_pos[:, 2] = Z.ravel()
